Updated sparky to working state.

+++ sparky/src/sparky-core/servo_angle_controller.cpp
@@ -9,21 +9,21 @@ ServoAngleController::ServoAngleController()
 }
 
 //
-ServoAngleController::ServoAngleController( const uint16_t n_channels, const std::string path, const bool connect ) :
+ServoAngleController::ServoAngleController( const uint8_t n_channels, const std::string path, const bool connect ) :
     ServoController( n_channels, path, connect )
 {
     ServoAngleController::init( 1, n_channels );
 } // ServoAngleController(const uint16_t, const std::string, const bool)
 
 //
-ServoAngleController::ServoAngleController( const uint16_t n_devices, const uint16_t n_channels_each, const std::string path, const bool connect ) :
+ServoAngleController::ServoAngleController( const uint8_t n_devices, const uint8_t n_channels_each, const std::string path, const bool connect ) :
     ServoController( n_devices, n_channels_each, path, connect )
 {
     ServoAngleController::init( n_devices, n_channels_each );
 } // ServoAngleController(const uint16_t, const uint16_t, const std::string, const bool)
 
 //
-ServoAngleController::ServoAngleController( const std::vector<uint16_t> n_device_channels, const std::string path, const bool connect ) :
+ServoAngleController::ServoAngleController( const std::vector<uint8_t> n_device_channels, const std::string path, const bool connect ) :
     ServoController( n_device_channels, path, connect )
 {
     ServoAngleController::init( n_device_channels );
@@ -70,13 +70,13 @@ ServoAngleController::~ServoAngleController()
 } // ~ServoAngleController()
 
 //
-bool ServoAngleController::isValidServoAngleTarget( const uint16_t channel, const double target ) const
+bool ServoAngleController::isValidServoAngleTarget( const uint8_t channel, const double target ) const
 {
     return isValidServoAngleTarget( 0, channel, target );
 } // isValidServoAngleTarget(const uint16_t, const double)
 
 //
-bool ServoAngleController::isValidServoAngleTarget( const uint16_t device, const uint16_t channel, const double target ) const
+bool ServoAngleController::isValidServoAngleTarget( const uint8_t device, const uint8_t channel, const double target ) const
 {
     if ( !isValidChannel( device, channel ) ) return false;
     return ( target >= getServoAngleMinLimit( device, channel ) ) && ( target
@@ -84,13 +84,13 @@ bool ServoAngleController::isValidServoAngleTarget( const uint16_t device, const
 } // isValidServoAngleTarget(const uint16_t, const uint16_t, const double)
 
 //
-double ServoAngleController::clipServoAngleTargetValue( const uint16_t channel, const double target ) const
+double ServoAngleController::clipServoAngleTargetValue( const uint8_t channel, const double target ) const
 {
     return clipServoAngleTargetValue( 0, channel, target );
 } // clipServoAngleTargetValue(const uint16_t, const double)
 
 //
-double ServoAngleController::clipServoAngleTargetValue( const uint16_t device, const uint16_t channel, const double target ) const
+double ServoAngleController::clipServoAngleTargetValue( const uint8_t device, const uint8_t channel, const double target ) const
 {
     if ( !isValidChannel( device, channel ) ) return 0.0;
 
@@ -102,13 +102,13 @@ double ServoAngleController::clipServoAngleTargetValue( const uint16_t device, c
 } // clipServoAngleTargetValue(const uint16_t, const uint16_t, const double)
 
 //
-double ServoAngleController::convertServoToAngle( const uint16_t channel, uint16_t pos ) const
+double ServoAngleController::convertServoToAngle( const uint8_t channel, uint16_t pos ) const
 {
     return convertServoToAngle( 0, channel, pos );
 } // convertServoToAngle(const uint16_t, uint16_t)
 
 //
-double ServoAngleController::convertServoToAngle( const uint16_t device, const uint16_t channel, uint16_t pos ) const
+double ServoAngleController::convertServoToAngle( const uint8_t device, const uint8_t channel, uint16_t pos ) const
 {
     if ( !isValidChannel( device, channel ) ) return 0.0;
 
@@ -124,13 +124,13 @@ double ServoAngleController::convertServoToAngle( const uint16_t device, const u
 } // convertServoToAngle(const uint16_t, const uint16_t, uint16_t)
 
 //
-uint16_t ServoAngleController::convertAngleToServo( const uint16_t channel, double angle ) const
+uint16_t ServoAngleController::convertAngleToServo( const uint8_t channel, double angle ) const
 {
     return convertAngleToServo( 0, channel, angle );
 } // convertAngleToServo(const uint16_t, double)
 
 //
-uint16_t ServoAngleController::convertAngleToServo( const uint16_t device, const uint16_t channel, double angle ) const
+uint16_t ServoAngleController::convertAngleToServo( const uint8_t device, const uint8_t channel, double angle ) const
 {
     if ( !isValidChannel( device, channel ) ) return 0;
 
@@ -146,13 +146,13 @@ uint16_t ServoAngleController::convertAngleToServo( const uint16_t device, const
 } // convertAngleToServo(const uint16_t, const uint16_t, double)
 
 //
-bool ServoAngleController::setServoAngleLimits( const uint16_t channel, ServoAngleLimits limits )
+bool ServoAngleController::setServoAngleLimits( const uint8_t channel, ServoAngleLimits limits )
 {
     return setServoAngleLimits( 0, channel, limits );
 } // setServoAngleLimits(const uint16_t, ServoAngleLimits)
 
 //
-bool ServoAngleController::setServoAngleLimits( const uint16_t device, const uint16_t channel, ServoAngleLimits limits )
+bool ServoAngleController::setServoAngleLimits( const uint8_t device, const uint8_t channel, ServoAngleLimits limits )
 {
     PRINT_DEBUG( "ServoAngleController::setServoAngleLimits( %u, %u, ServoAngleLimits )\n", device, channel );
     //if ((!isConnected()) || (!isValidChannel(device, channel)))
@@ -166,13 +166,13 @@ bool ServoAngleController::setServoAngleLimits( const uint16_t device, const uin
 } // setServoAngleLimits(const uint16_t, const uint16_t, ServoAngleLimits)
 
 //
-bool ServoAngleController::setServoAngleLimits( const uint16_t channel, ServoAnglePair limit1, ServoAnglePair limit2 )
+bool ServoAngleController::setServoAngleLimits( const uint8_t channel, ServoAnglePair limit1, ServoAnglePair limit2 )
 {
     return setServoAngleLimits( 0, channel, limit1, limit2 );
 } // setServoAngleLimits(const uint16_t, ServoAngle, ServoAngle)
 
 //
-bool ServoAngleController::setServoAngleLimits( const uint16_t device, const uint16_t channel, ServoAnglePair limit1, ServoAnglePair limit2 )
+bool ServoAngleController::setServoAngleLimits( const uint8_t device, const uint8_t channel, ServoAnglePair limit1, ServoAnglePair limit2 )
 {
     //if ((!isConnected()) || (!isValidChannel(device, channel)))
     if ( !isValidChannel( device, channel ) ) return false;
@@ -182,7 +182,7 @@ bool ServoAngleController::setServoAngleLimits( const uint16_t device, const uin
 } // setServoAngleLimits(const uint16_t, const uint16_t, ServoAngle, ServoAngle)
 
 // commands an individual servo motor to move to a target position
-bool ServoAngleController::setServoAngleTarget( const uint16_t channel, double target )
+bool ServoAngleController::setServoAngleTarget( const uint8_t channel, double target )
 {
     //if ((!isConnected()) || (channel >= getNumChannels(0)) || (!isValidServoAngleTarget(channel, target)))
     if ( ( channel >= getNumChannels( 0 ) ) || ( !isValidServoAngleTarget( channel, target ) ) ) return false;
@@ -190,7 +190,7 @@ bool ServoAngleController::setServoAngleTarget( const uint16_t channel, double t
 } // setServoAngleTarget(const uint16_t, double)
 
 // commands an individual servo motor to move to a target position
-bool ServoAngleController::setServoAngleTarget( const uint16_t device, const uint16_t channel, double target )
+bool ServoAngleController::setServoAngleTarget( const uint8_t device, const uint8_t channel, double target )
 {
     //if ((!isConnected()) || (!isValidChannel(device, channel)) || (!isValidServoAngleTarget(device, channel, target)))
     if ( !isValidChannel( device, channel ) )
@@ -213,7 +213,7 @@ bool ServoAngleController::setServoAngleTarget( const uint16_t device, const uin
 } // setServoAngleTarget(const uint16_t, const uint16_t, double)
 
 // commands an individual servo motor to move at the parameterized speed
-bool ServoAngleController::setServoAngleSpeed( const uint16_t channel, double speed )
+bool ServoAngleController::setServoAngleSpeed( const uint8_t channel, double speed )
 {
     //if (!isConnected() || (!isValidChannel(channel)))
     if ( !isValidChannel( channel ) ) return false;
@@ -221,7 +221,7 @@ bool ServoAngleController::setServoAngleSpeed( const uint16_t channel, double sp
 } // setServoAngleSpeed(const uint16_t, double)
 
 // commands an individual servo motor to move at the parameterized speed
-bool ServoAngleController::setServoAngleSpeed( const uint16_t device, const uint16_t channel, double speed )
+bool ServoAngleController::setServoAngleSpeed( const uint8_t device, const uint8_t channel, double speed )
 {
     //if (!isConnected() || (!isValidChannel(device, channel)))
     if ( !isValidChannel( device, channel ) ) return false;
@@ -229,7 +229,7 @@ bool ServoAngleController::setServoAngleSpeed( const uint16_t device, const uint
 } // setServoAngleSpeed(const uint16_t, const uint16_t, double)
 
 // commands an individual servo motor to move at the parameterized acceleration
-bool ServoAngleController::setServoAngleAcceleration( const uint16_t channel, double accel )
+bool ServoAngleController::setServoAngleAcceleration( const uint8_t channel, double accel )
 {
     //if (!isConnected() || (!isValidChannel(channel)))
     if ( !isValidChannel( channel ) ) return false;
@@ -237,7 +237,7 @@ bool ServoAngleController::setServoAngleAcceleration( const uint16_t channel, do
 } // setServoAngleAcceleration(const uint16_t, double)
 
 // commands an individual servo motor to move at the parameterized acceleration
-bool ServoAngleController::setServoAngleAcceleration( const uint16_t device, const uint16_t channel, double accel )
+bool ServoAngleController::setServoAngleAcceleration( const uint8_t device, const uint8_t channel, double accel )
 {
     //if (!isConnected() || (!isValidChannel(device, channel)))
     if ( !isValidChannel( device, channel ) ) return false;
@@ -245,13 +245,13 @@ bool ServoAngleController::setServoAngleAcceleration( const uint16_t device, con
 } // setServoAcceleration(const uint16_t, const uint16_t, double)
 
 //
-ServoAngleLimits ServoAngleController::getServoAngleLimits( const uint16_t channel ) const
+ServoAngleLimits ServoAngleController::getServoAngleLimits( const uint8_t channel ) const
 {
     return getServoAngleLimits( 0, channel );
 } // getServoAngleLimits(const uint16_t)
 
 //
-ServoAngleLimits ServoAngleController::getServoAngleLimits( const uint16_t device, const uint16_t channel ) const
+ServoAngleLimits ServoAngleController::getServoAngleLimits( const uint8_t device, const uint8_t channel ) const
 {
     //if ((!isConnected()) || (!isValidChannel(device, channel)))
     if ( !isValidChannel( device, channel ) ) return ServoAngleLimits( ServoAnglePair( 0, 0.0 ), ServoAnglePair( 0, 0.0 ) );
@@ -259,13 +259,13 @@ ServoAngleLimits ServoAngleController::getServoAngleLimits( const uint16_t devic
 } // getServoAngleLimits(const uint16_t, const uint16_t)
 
 //
-ServoAnglePair ServoAngleController::getServoAngleMinLimitPair( const uint16_t channel ) const
+ServoAnglePair ServoAngleController::getServoAngleMinLimitPair( const uint8_t channel ) const
 {
     return getServoAngleMinLimitPair( 0, channel );
 } // getServoAngleMinLimitPair(const uint16_t)
 
 //
-ServoAnglePair ServoAngleController::getServoAngleMinLimitPair( const uint16_t device, const uint16_t channel ) const
+ServoAnglePair ServoAngleController::getServoAngleMinLimitPair( const uint8_t device, const uint8_t channel ) const
 {
     ServoAngleLimits limits = getServoAngleLimits( device, channel );
     return ( limits.first.second < limits.second.second ) ? limits.first
@@ -273,13 +273,13 @@ ServoAnglePair ServoAngleController::getServoAngleMinLimitPair( const uint16_t d
 } // getServoAngleMinLimitPair(const uint16_t, const uint16_t)
 
 //
-ServoAnglePair ServoAngleController::getServoAngleMaxLimitPair( const uint16_t channel ) const
+ServoAnglePair ServoAngleController::getServoAngleMaxLimitPair( const uint8_t channel ) const
 {
     return getServoAngleMaxLimitPair( 0, channel );
 } // getServoAngleMaxLimitPair(const uint16_t)
 
 //
-ServoAnglePair ServoAngleController::getServoAngleMaxLimitPair( const uint16_t device, const uint16_t channel ) const
+ServoAnglePair ServoAngleController::getServoAngleMaxLimitPair( const uint8_t device, const uint8_t channel ) const
 {
     ServoAngleLimits limits = getServoAngleLimits( device, channel );
     return ( limits.first.second > limits.second.second ) ? limits.first
@@ -287,73 +287,73 @@ ServoAnglePair ServoAngleController::getServoAngleMaxLimitPair( const uint16_t d
 } // getServoAngleMaxLimitPair(const uint16_t, const uint16_t)
 
 //
-double ServoAngleController::getServoAngleMinLimit( const uint16_t channel ) const
+double ServoAngleController::getServoAngleMinLimit( const uint8_t channel ) const
 {
     return getServoAngleMinLimit( 0, channel );
 } // getServoAngleMinLimit(const uint16_t)
 
 //
-double ServoAngleController::getServoAngleMinLimit( const uint16_t device, const uint16_t channel ) const
+double ServoAngleController::getServoAngleMinLimit( const uint8_t device, const uint8_t channel ) const
 {
     return getServoAngleMinLimitPair( device, channel ).second;
 } // getServoAngleMinLimit(const uint16_t, const uint16_t)
 
 //
-double ServoAngleController::getServoAngleMaxLimit( const uint16_t channel ) const
+double ServoAngleController::getServoAngleMaxLimit( const uint8_t channel ) const
 {
     return getServoAngleMinLimit( 0, channel );
 } // getServoAngleMaxLimit(const uint16_t)
 
 //
-double ServoAngleController::getServoAngleMaxLimit( const uint16_t device, const uint16_t channel ) const
+double ServoAngleController::getServoAngleMaxLimit( const uint8_t device, const uint8_t channel ) const
 {
     return getServoAngleMaxLimitPair( device, channel ).second;
 } // getServoAngleMaxLimit(const uint16_t, const uint16_t)
 
 //
-double ServoAngleController::getServoAngleTarget( const uint16_t channel )
+double ServoAngleController::getServoAngleTarget( const uint8_t channel )
 {
     return getServoAngleTarget( 0, channel );
 } // getServoAngleTarget(const uint16_t)
 
 //
-double ServoAngleController::getServoAngleTarget( const uint16_t device, const uint16_t channel )
+double ServoAngleController::getServoAngleTarget( const uint8_t device, const uint8_t channel )
 {
     return convertServoToAngle( device, channel, getServoTarget( device, channel ) );
 } // getServoAngleTarget(const uint16_t, const uint16_t)
 
 //
-double ServoAngleController::getServoAngleSpeed( const uint16_t channel )
+double ServoAngleController::getServoAngleSpeed( const uint8_t channel )
 {
     return getServoAngleSpeed( 0, channel );
 } // getServoAngleSpeed(const uint16_t)
 
 //
-double ServoAngleController::getServoAngleSpeed( const uint16_t device, const uint16_t channel )
+double ServoAngleController::getServoAngleSpeed( const uint8_t device, const uint8_t channel )
 {
     return convertServoToAngle( device, channel, getServoSpeed( device, channel ) );
 } // getServoAngleSpeed(const uint16_t, const uint16_t)
 
 //
-double ServoAngleController::getServoAngleAcceleration( const uint16_t channel )
+double ServoAngleController::getServoAngleAcceleration( const uint8_t channel )
 {
     return getServoAngleAcceleration( 0, channel );
 } // getServoAngleAcceleration(const uint16_t)
 
 //
-double ServoAngleController::getServoAngleAcceleration( const uint16_t device, const uint16_t channel )
+double ServoAngleController::getServoAngleAcceleration( const uint8_t device, const uint8_t channel )
 {
     return convertServoToAngle( device, channel, getServoAcceleration( device, channel ) );
 } // getServoAngleAcceleration(const uint16_t, const uint16_t)
 
 //
-double ServoAngleController::getServoAnglePosition( const uint16_t channel )
+double ServoAngleController::getServoAnglePosition( const uint8_t channel )
 {
     return convertServoToAngle( channel, getServoPosition( channel ) );
 } // getServoAnglePosition(const uint16_t)
 
 //
-double ServoAngleController::getServoAnglePosition( const uint16_t device, const uint16_t channel )
+double ServoAngleController::getServoAnglePosition( const uint8_t device, const uint8_t channel )
 {
     return convertServoToAngle( device, channel, getServoPosition( device, channel ) );
 } // getServoAnglePosition(const uint16_t, const uint16_t)
@@ -372,20 +372,20 @@ ServoAngleController& ServoAngleController::operator =( const ServoAngleControll
 bool ServoAngleController::init()
 {
     PRINT_DEBUG( "ServoAngleController::init()\n" );
-    uint16_t n_devices = getNumDevices();
+    uint8_t n_devices = getNumDevices();
     if ( n_devices == 0 )
     {
         PRINT_WARN( "Init failed because n_devices = 0\n" );
         return false;
     }
-    std::vector<uint16_t> n_device_channels( n_devices );
+    std::vector<uint8_t> n_device_channels( n_devices );
     for ( int i = 0; i < n_devices; ++i )
         n_device_channels[i] = getNumChannels( i );
     return init( n_device_channels );
 } // init()
 
 //
-bool ServoAngleController::init( const uint16_t n_devices, const uint16_t n_channels_each )
+bool ServoAngleController::init( const uint8_t n_devices, const uint8_t n_channels_each )
 {
     PRINT_DEBUG( "ServoAngleController::init( %u %u )\n", n_devices, n_channels_each );
     if ( ( n_devices == 0 ) || ( n_channels_each == 0 ) )
@@ -400,7 +400,7 @@ bool ServoAngleController::init( const uint16_t n_devices, const uint16_t n_chan
 } // init(const uint16_t, const uint16_t)
 
 //
-bool ServoAngleController::init( const std::vector<uint16_t> n_device_channels )
+bool ServoAngleController::init( const std::vector<uint8_t> n_device_channels )
 {
     PRINT_DEBUG( "ServoAngleController::init( std::vector( %zu ) )\n", n_device_channels.size() );
     unsigned int n_devices = n_device_channels.size();
@@ -436,8 +436,8 @@ void ServoAngleController::initFromYaml( YAML::Node const & node )
         YAML::Node const & servos_node = first_node["servos"];
         for ( int i = 0; i < servos_node.size(); ++i )
         {
-            uint16_t const & device = this->servos_[0][i].device_;
-            uint16_t const & channel = this->servos_[0][i].channel_;
+            uint8_t const & device = this->servos_[0][i].device_;
+            uint8_t const & channel = this->servos_[0][i].channel_;
 
             ServoAngleLimits servo_angle_limits;
             servos_node[i]["angle_limits"] >> servo_angle_limits;
